+++ module/testing/testing-full.js
@@ -17,7 +17,7 @@ export function registerFullTests(quench) {
                     <?xml version="1.0" encoding="UTF-16"?>
                     <CHARACTER version="6.0" TEMPLATE="builtIn.Superheroic.hdt">
                     <BASIC_CONFIGURATION BASE_POINTS="200" DISAD_POINTS="150" EXPERIENCE="0" RULES="Default" />
-                    <CHARACTER_INFO CHARACTER_NAME="5e superhero" ALTERNATE_IDENTITIES="" PLAYER_NAME="" HEIGHT="78.74015748031496" WEIGHT="220.46224760379584" HAIR_COLOR="Brown" EYE_COLOR="Brown" CAMPAIGN_NAME="" GENRE="" GM="">
+                    <CHARACTER_INFO CHARACTER_NAME="5e superhero simple" ALTERNATE_IDENTITIES="" PLAYER_NAME="" HEIGHT="78.74015748031496" WEIGHT="220.46224760379584" HAIR_COLOR="Brown" EYE_COLOR="Brown" CAMPAIGN_NAME="" GENRE="" GM="">
                         <BACKGROUND />
                         <PERSONALITY />
                         <QUOTE />
@@ -102,7 +102,7 @@ export function registerFullTests(quench) {
                 it("name", async function () {
                     console.log("name")
                     await actor.uploadFromXml(contents)
-                    assert.equal(actor.name, "5e superhero");
+                    assert.equal(actor.name, "5e superhero simple");
                 });
 
                 it("str.max", async function () {
@@ -1227,7 +1227,7 @@ export function registerFullTests(quench) {
                         type: 'pc',
                     }, { temporary: true });
                     await actor.uploadFromXml(contents)
-                    assert.equal(actor.system.characteristics.running.max, 8);
+                    assert.equal(actor.system.characteristics.running.max, 8);  //8 + 5 (Active Effect)
                 });
                 it("running.realCost", async function () {
                     const actor = new HeroSystem6eActor({
@@ -1278,7 +1278,7 @@ export function registerFullTests(quench) {
                         type: 'pc',
                     }, { temporary: true });
                     await actor.uploadFromXml(contents)
-                    assert.equal(actor.system.realCost, 675);
+                    assert.equal(actor.system.realCost, 732);
                 });
 
             });